requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-cors-headers==3.8.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
packaging==21.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.3
uritemplate==3.0.1
urllib3==1.26.6
